feature: query latest sensor data from TSDB.
