+++ package.json
@@ -16,6 +16,7 @@
     "@mui/x-charts": "^7.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.0",
     "three": "^0.164.1",
     "three-pathfinding": "^1.2.0",
     "uuid": "^9.0.1"
