+++ server/nginx/nginx.conf
@@ -11,8 +11,8 @@ server {
         proxy_set_header X-Real-IP $remote_addr;
     }
 
-    location /triangle-localization-simulator {
-        proxy_pass http://client:5173/triangle-localization-simulator;
+    location /triangle-localization-simulator/ {
+        proxy_pass http://client:5173/triangle-localization-simulator/;
         proxy_set_header Host $host;
         proxy_set_header X-Real-IP $remote_addr;
     }
